ch3. ispolzuem funcii

+++ js/script6.js
@@ -22,4 +22,5 @@ const num = 12.2;
 console.log(Math.round(num));
 
 const test = "12.2px";
-console.log(parseInt(test));
+// console.log(parseInt(test));
+console.log(parseFloat(test));
